requirements.txt
numpy==1.26.4
onnx
onnxsim
opencv-python
pyyaml
onnxruntime
